fix(servicios): cantidad de servicio entera con tope de negocio (Fase 1) La cantidad de pedidos y consumos de servicio aceptaba decimales y no tenía límite superior, permitiendo registros sin sentido de negocio (p. ej. "Lavandería" cantidad 41). Se estandariza a un entero de unidades con validación de tope en el backend como autoridad final. Cambios: - cantidad pasa de BigDecimal a Integer en pedido del cliente y consumo de   recepción (entidades, DTOs de request/response, mappers y payload WebSocket). - Bean Validation: @Min(1) + @Max(99) como tope duro de cordura en los request. - Tope de negocio CANTIDAD_MAXIMA_DEFAULT = 50 validado en   PedidoServicioServiceImpl.crear(), con mensaje legible que incluye el tipo y el   tope (default acordado con frontend para cantidadMaxima null). - Subtotales recalculados con BigDecimal.valueOf(cantidad); el dinero   (precioAplicado, subtotal, costoBase) sigue siendo BigDecimal. - Migración V28: pedidos_servicio.cantidad y servicios.cantidad de NUMERIC(10,2)   a INTEGER (USING ROUND, datos existentes son enteros .00).

+++ src/main/java/com/sanfrancisco/api/modules/servicios/dto/response/PedidoServicioResponse.java
@@ -10,7 +10,7 @@ public record PedidoServicioResponse(
         Integer tipoServicioId,
         String tipoServicioNombre,
         BigDecimal costoBase,
-        BigDecimal cantidad,
+        Integer cantidad,
         BigDecimal subtotalEstimado,
         String observaciones,
         EstadoPedidoServicio estado,
